Be sure to return indices astype int

+++ visbrain/sleep/interface/ui_elements/ui_detection.py
@@ -154,7 +154,7 @@ class UiDetection(object):
                 return idx
             # Check indices shape and format to (n_events, 2) :
             if (idx.ndim == 2) and (idx.shape[1] == 2):  # (n_events, 2)
-                return idx
+                return idx.astype(int)
             elif idx.ndim == 1:  # 1d vector
                 if idx.dtype == bool:  # boolean array
                     assert len(idx) == len(data)
